requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
Django==5.1.4
firebase-admin==6.6.0
gcloud==0.18.3
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
joblib==1.4.2
jwcrypto==1.5.6
msgpack==1.1.0
numpy==2.2.1
oauth2client==4.1.3
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
pyparsing==3.2.1
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-jwt==4.1.0
pytz==2024.2
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.20

yfinance~=0.2.52
channels~=4.2.0